update and remove caption

+++ backend/server.js
@@ -140,13 +140,17 @@ app.delete("/api/photodb", async (req, res) => {
   const imageData = req.body;
   const resObj = {
     imageRemoved: false,
+    captionRemoved: false,
   };
 
   let photo = await photosDB.find({ _id: imageData.imageID });
 
-  if (photo.length > 0) {
+  if (photo.length > 0 && imageData.removeCaption !== true) {
     resObj.imageRemoved = true;
     photosDB.remove({ _id: imageData.imageID });
+  } else if (photo.length > 0 && imageData.removeCaption === true) {
+    photosDB.update({ _id: imageData.imageID }, { $set: { caption: "" } });
+    resObj.captionRemoved = true;
   }
 
   res.json(resObj);
@@ -172,8 +176,7 @@ app.post("/api/photoInfo", async (req, res) => {
   const userID = req.body.userID;
   const _id = req.body._id;
   const caption = req.body.caption;
-  console.log(caption);
-  console.log(userID);
+
   const resObj = {
     username: "",
     caption: caption,
